[SPARK-51604][PYTHON][TESTS] Split `test_connect_session`
### What changes were proposed in this pull request?
Split `test_connect_session`

### Why are the changes needed?
the test is somewhat flaky, e.g. occasional timeout  https://github.com/apache/spark/actions/runs/14044733934/attempts/1

split it so that we can narrow the scope

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
updated ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #50397 from zhengruifeng/py_test_session_split.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
